PORT TO STORE CORPUS IN BERKELEYDB DATABASES
Bayes.pm:

The $self->{matrix__} hash is now a collection of tied hashes
to BerkeleyDB databases named table.db in each of the corpus
bucket subdirectories.  The set_value_ and get_value_ accessors
have been modified to access the database.  load_word_matrix_
and load_bucket_ now load the bucket information from the database
in concurrent mode.

prefork, forked and postfork handling closing and opening database
connections around forks to ensure that there are no threading
problems with the database.

close_database__ can be called to clean up the connection to the
database at any time.

Many API functions have been modified internally to use the new
structure.  The external APIs have not changed.  get_bucket_word_list
is currently not implemented.

load_bucket_ does automatic upgrade from the old flat file style
of corpus to the database.


HTML.pm:

Added a note that since get_bucket_word_list isn't working it is
not possible to view the words in a bucket.


Module.pm:

Added description and base implementation of the new postfork()
method that is called on all modules when a fork has occurred and
in the parent process.   This is the parent equivalent of forked().


Loader.pm:

The forker is modified to call postfork() in the parent process
after a successful fork.