ADDITION OF FORCE_FORK PARAMETER FOR PROXIES
Proxy.pm:

Handle the force_fork parameter.  If it is 1 then when we get
a connection from a client we fork, if it is 0 then we handle
the connection in line and queue up connections using the standard
TCP mechanism up to the maximum backlog for the platform.


POP3.pm:
NNTP.pm:

Default to not forking


SMTP.pm:

Default to forking