USING PERLTRAY TO BUILD A POPFILE.EXE FOR WINDOWS
Makefile:

Suitable targets to use PerlTray to created popfile.exe and
popfileb.exe.


POPFile/Loader.pm:

Added version() accessor for version string and an option to
service so that it can be called from an external poller.


popfilew.pl:

Equivalent of popfile.pl which uses PerlTray to show an icon
in the system tray on Windows.   This is designed to be turned
into popfile.exe by PerlTray.