MAKE POPFILE BE A TRUE EXE ON WINDOWS ADDED ATTACHMENT PSEUDOWORDS
Makefile:

At rules for making a variety of POPFile executables:

popfile.exe         Loader that picks one of the following four to run:
    popfileif.exe       Run POPFile in foreground with a tray icon
    popfileib.exe       Run POPFile in background with a tray icon
    popfilef.exe        Run POPFile in the foreground
    popfileb.exe        Run POPFile in the background

Each executable is built with either PerlTray (for those with system
tray icons) or PerlApp (for those with no icon).

popfile-tray.pl:

Equivalent of popfile.pl but with hooks for use by PerlTray for
showing the icon and handling the popup menu.

popfile-windows.pl:

This is the small wrapper that becomes popfile.exe and decides which
version of POPFile to load.

languages/English.msg:

Added the message for the console output option and to warn the user
that the change isn't going to take affect until the next restart.

Platform/MSWin32.pm:

This just contains code now for handling the options windows_trayicon
and windows_console.

Classifier/Bayes.pm:

Removed the second blessing of ourselves because it was not needed.

Classifier/MailParse.pm:

Added pseudowords for attachment names and attachment extensions via a
patch that was submitted.

POPFile/Loader.pm:

New public function module_config() allows a user of the loader to access
configuration items without knowing the internals of Loader.

popfilew.pl:

Removed.  This file is replaced by popfile-windows.pl.