Classifier/Bayes.pm:
Factor the tieing and untieing of the BerkeleyDB hash into
helper function tie_bucket__ and untie_bucket__ to remove
duplicated code.

When doing a bucket upgrade untie and retie the BerkeleyDB
hash to ensure that the database has been closed between
creation and verification.

When tieing to an empty hash fix a bug where the TOTAL and
UNIQUE scores were not getting correctly initialized.

Make create_bucket use the tie_bucket__ and untie_bucket__
functions which fixes an odd situation where a db gets
tied twice.