Optimize the performance of the following functions:
1. Display of the word list in the Single Message View
2. Message reclassification

Classifier/MailParse.pm:

start_parse and parse_file have an optional 'reset' parameter that specifies
whether the word list from a previous parse is reset of not.  This is used
by reclassification so that multiple messages can be parsed and a single
word list created.

Classifier/Bayes.pm:

Speed up db_update_cache__ doing a single query to get the bucket word
counts instead of one query per bucket.

Modify add_words_to_bucket__ so that it works in the most efficient manner
possible by getting the list of wordids, then pulling the list of values
and then using that data to update the database within a transaction.

Optimized the generation of the table of word scores/probabilities/frequences
so that it needs no further database access after the classification since
all the data is needs is in temporary %matrix hash.