Optimize the performance the buckets page by caching the value of the unique word count per bucket.
Fix bug where we were able to create a new bucket with the same name
as another bucket.

Classifier/Bayes.pm:

db_updatecache__ now keeps track of the unique word count per bucket and
the get_unique_word_count API returns the cached value.

create_bucket returns an error if the bucket already exists.

When creating the database schema, check the syntax of the SQL file so
that we remove 'alter' commands when trying to create SQLite since it is
not supported and not needed.

Make sure that we clear the cached count values in case they go to
zero.

UI/HTML.pm:

Use is_bucket and is_pseudo_bucket to check for bucket existence before
calling create_bucket.  Don't output a false negative count for pseudo
buckets.

tests/TestBayes.tst:

Add test for trying to create a bucket with an existing name.  Add tests
to ensure that when a bucket is deleted it has gone.