popfile-tray.pl:
Implement the PerlTray Shutdown callback so that we get correct
termination of POPFile is it is running in the system tray and
Windows shuts down (or the current user logs off).

popfile-service.pl:

(EXPERIMENTAL) POPFile wrapped up as a Windows service using
ActiveState's PerlSvc.  This file is turning into popfile-service.exe
by instructions on the Makefile.

Makefile:

Add instructions to build the Windows service version of POPFile.