Added SSL support for POP3 over SSL (and other proxies if needed).
Proxy/Proxy.pm:

verify_connected_ can now establish SSL connections if the $ssl
parameter is set to 1.

Proxy/POP3.pm:

Take Sam's syntax for doing APOP on the USER command and expand it to be
a generic "options" syntax so that the USER settings in the POP3 client
are not in the form

    host:port:username:options

where port and options are optional.

Valid options are now

   apop         Require APOP authentication
   ssl          Require SSL connection from POPFile to server

it's possible to use both as follows:

    host:port:username:apop,ssl