Had to change the separator in the config vars from tab to '-->'. When using the Advanced page, the tabs were transformed to spaces. I don't know whether html.pm was responsible, but it made me aware of the fact that a naughty browser could also break that in a horrible way. The separator now has its own class variable $cfg_separator.
Improved validate_item():
- we now check whether the bucket->folder mapping uses any folder twice,
  since we reclassify_on_move relies on a unique folder->bucket mapping.
- If two watched folder entries are set to the same folder, one of them
  is removed. So we now have an (admittedly awkward) way to remove watched
  folders.

Added error checking to say() and get_response(). If print or slurp_
fail, we die with a warning and set $self->{imap__} to '' which triggers
reconnection and new login the next time through service() which will catch
the exception.

Added a control for the debug level to the UI. You now no longer need to
restart POPFile to adjust the log noise.