* Service now sets local $SIG{PIPE} to 'IGNORE' which will hopefully fix the problem Naoki is seeing when his machine wakes up again.
* reclassify_on_move now tries to respect messages that are
completely new, so they can be classified after it's done.

* Made a couple of functions private to the IMAP class. The
only public functions are now those that are inherited or
registered handlers plus the IMAP utility functions that
do not rely on internal data.

* Added another configuration variable along with a
corresponding UI widget: I made it optional whether we should
"CLOSE" folders (and therefore trigger an EXPUNGE) when we leave
a folder. This will only happen if we moved something out of
the folder before.