Refactored check_for_new_messages() and reclassify_on_move__(): Both subs are no longer with us.
Instead, we now have:
-scan_folders
which scans all of our folders and decides what to do with new messages.
If they should be classified, classify_message() is called.
If they should be reclassifed, reclassify_message() is called.

This should eliminate the race condition we had previously since both classify and reclassify get to look at the same list
of messages.