For the 0.22 release the system tray icon will use 'localhost' instead of '127.0.0.1' to display the User Interface, so a new library !define has been created to replace the hard-coded references to '127.0.0.1'.
Improve the handling of conditional compilation by introducing a new
"!define" for the main installer script to simplify the conditional
expressions.

Provide support for "PFI Diagnostic" and "Silent Shutdown" utilities.

New functions to check if any of the PFI utilities are running and
to find the location of the SQLite database.

Use new customisable shutdown message boxes.