For the 0.22 release the system tray icon will use 'localhost' instead of '127.0.0.1' to display the User Interface, so the wizard now uses a library !define instead of hard-coded references to '127.0.0.1' for UI.
Provide more [un]installation progress messages, especially for those
operations that may take a few seconds.

If SQLite is not used for the SQL database, do not create a shortcut
to the SQLite utility in the 'User Data' folder.

Create a backup copy of the existing SQLite database if it does not
contain message history (using a simple test on the messages folder
instead of examining the database itself).

Update the 'pfi-run.bat' batch file to use the new Message Capture
utility (if it can be found) when running in debug mode.

When the old SQLite database is being converted, use the new Message
Capture utility to display the progress reports.

Use the new customised "manual shutdown" message boxes.

Uninstaller actions rearranged into a more logical order with better
handling of some of the special cases.

Make uninstaller able to delete a non-default 'messages' folder
(but ask user for permission first).

[Still need to improve the various "start POPFile" cases]