This is the reason we have test suites.
I got the TestPOP3 suite working with the new History mechanism and
MQ handling and it showed up a bug that has been introduced in the
TOP handling.  If toptoo was enabled and a TOP was sent then the messages
would end up with the wrong XPL link in it.  That would have been very
hard to track down in the field.

This necessitates a change in the signature to classify_and_modify.

tests/TestPOP3.tst:

Change from the old way we were accessing the history to the actually
use the History module.  Put in lots of calls to MQ::service() to handle
all the child messages.

Complete POP3 test suite passes at this point.

Classifier/Bayes.pm:

classify_and_modify gets an additional parameter $slot which is set
when $class is set.  If $class is set then don't reserve a new history
slot because the assumption is that the message has already been classified.

Proxy/POP3.pm
Proxy/SMTP.pm
Proxy/NNTP.pm:

Update calls to classify_and_modify for the new signature.

tests/TestBayes.tst:

Update tests for classify_and_modify.