Various bits of pre-v0.22.0 clean up.
POPFile/Logger.pm:

Change the behaviour of Logger so that it sets the debug_filename__ parameter
in the start method and wont log anything until that's been called.  This
means that we wont log until the logdir has been correctly set.

UI/HTML.pm:

Added a new parameter html_strict_templates which defaults to 0 and if set
to 1 sets the 'die_on_bad_params' parameter for HTML::Template.

POPFile/History.pm
tests/TestHistory.tst:

Add code to clean up empty history directories when files are deleted.
Add tests for that functionality.

UI/HTML.pm
POPFile/Loader.pm:

"Fixed" problem with multiple clicks in Linux causing the program to
terminate with either SIGPIPE or silently.   After extensive debugging
the problem is that an unhandled SIGPIPE causes perl itself to terminate
without terminating POPFile.  Set SIGPIPE to IGNORE so that we ignore
it completely.   'perlipc' recommends ignoring SIGPIPE but it seems a bit odd
just to ignore it completely.  *sigh*