Various pre-v0.22.0 release things that are all testing related.
languages/English.msg
tests/TestPOP3.tst
Proxy/POP3.pm:

Set an appropriate error message for unconfigured transparent proxying and
add tests for this functionality.  Change strings to mention transparent
proxying.

languages/Nihongo.msg
languages/English.msg
skins/default/common-bottom.thtml:

Make the location of the FAQ configurable on a per language basis.  Make a
link to the standard FAQ for all languages except Japanese where it goes
to the localized FAQ.

Also change "Additional Documentation" to "Documentation" in English.

UI/HTML.pm:

Fix a bug in localization that caused the Configuration page to be half
localized when switching languages.   Add new localize_template__ method
so that we can call it when the template is loaded and also when the language
is reloaded.

UI/HTML.pm:

Fix obscure bug to do with language loading that showed up if you switch to
Arabic (or another right-to-left language) and then back to a non-English
left-to-right language: you'd get right-to-left French for example.  Now
reload English before loading the next non-English language so that we get
the default parameters for languages.