Support two variants: 'integrated' (for bundling with the installer) and 'stand-alone' for use with 0.21.0 or a later release. The 'integrated' one does not need to include the sqlite.exe utility because the installer will install a compatible version of sqlite.exe.
Introduce a new run time option (/REGISTRY) to make it easier for the
installer to create a suitable Start Menu shortcut for the utility.

Make the report more user-friendly by automatically scrolling the output
to reveal the name and location of the POPFile SQLite database which was
examined (previously this line scrolled out of sight when the timestamp
was added at the end of the report).