More v0.23.0 work; mostly around the test suite and the ability to change passwords.
*** NOTE: Had to run 'dos2unix' because a number of files had CRLF line
    endings.  POPFile standard is Unix-style LF only line endings on
    files.

*** tests/TestBayes.tst and Classifier/Bayes.pm:

Add tests for APIs that had no tests (and fix the bugs that the tests
showed up!).  All APIs now have tests; the new tests cover:

    add_account
    create_user
    get_accounts
    get_bucket_id
    get_bucket_name
    get_user_id
    get_user_list
    get_user_parameter
    get_user_parameter_from_id
    get_user_parameter_list
    remove_account
    remove_user
    set_user_parameter_from_id

*** Classifier/Bayes.pm, tests/TestBayes.tst, UI/HTML.pm

create_user now also creates a random password for the user and the
password is displayed when the user is created in the UI.

*** Classifier/Bayes.pm, tests/TestBayes.tst, UI/HTML.pm,
    skins/default/configuration-bar.thtml

Added the ability to change password for a user.  This is done in the
configuration bar and added new APIs and tests (validate_password and
set_password) for a user.

*** POPFile/API.pm:

Add external access to the following new APIs:

    add_account
    create_user
    get_user_id
    get_user_id_from_session
    get_user_parameter
    get_user_parameter_from_id
    get_user_parameter_list
    remove_account
    remove_user
    set_user_parameter_from_id

*** tests/TestHTTP.tst:

Remove redirection tests since this is no longer handled by the HTTP
module.

*** Current state of the test suite:

    TestBayesScript         fail
    TestBayes                       PASS
    TestConfiguration       fail
    TestHistory             fail
    TestHTML                fail (horribly)
    TestHTTP                        PASS
    TestIMAP                        PASS
    TestInsertScript        fail
    TestLogger                      PASS
    TestMailParse           fail
    TestModule                      PASS
    TestMQ                          PASS
    TestMutex                       PASS
    TestPipeScript                  PASS
    TestPOP3                fail
    TestProxy                       PASS
    TestWordMangle                  PASS
    TestXMLRPC              fail

*** TODO

tests/TestModule.tst needs updating for the user_*_config_ APIs that
have been added.

Look into why we get bogus complaints about the primary key not being
unique when we do a database upgrade.  I think some of these are
caused by the fact that popfile.sql contains some INSERT statements
and we are getting double inserts.  Perhaps the upgrade should be
using UPDATE instead.