More v0.23.0 work.  In this case around getting more parts of the test suite to work.
*** POPFile/Database.pm:

Replace INSERT with INSERT OR IGNORE statements.  This surpresses the
errors we were seeing with duplicate primary keys caused by the
insertion of data from the schema file, followed by the insertion of
data.  The use of IGNORE also means that data inserted by the schema
file will override old data.

Making this change also causes the following tests to now pass:
TestBayesScript, TestInsertScript.

*** tests/TestConfiguration.tst

Update tests for new global parameter GLOBAL_single_user.  Fix a bug
that was causing the command-line parser to erroneously complain about
an empty -- command-line option.  Changed the default options for
Getopt::Long so that they are inline with the usage in
POPFile/Loader.pm.TestConfiguration test suite now passes.

*** POPFile/History.pm, tests/TestHistory.tst

Fix minor oddity in History where the query used for searches had a LF
in the middle of it (this was harmless but ugly).  Update test suite
so that start_query now gets a valid session.  TestHistory suite now
passes.

*** Current state of the test suite:

    TestBayesScript                 PASS
    TestBayes                       PASS
    TestConfiguration               PASS
    TestHistory                     PASS
    TestHTML                fail (horribly)
    TestHTTP                        PASS
    TestIMAP                        PASS
    TestInsertScript                PASS
    TestLogger                      PASS
    TestMailParse           fail
    TestModule                      PASS
    TestMQ                          PASS
    TestMutex                       PASS
    TestPipeScript                  PASS
    TestPOP3                fail
    TestProxy                       PASS
    TestWordMangle                  PASS
    TestXMLRPC              fail

TODO

Why is there no TestDatabase?

TestMailParse is failing on my machine because accented characters are
not being recognized as part of the [:alpha:] character class.  I have
not fully understood why yet.  Is any one else seeing this?  Try
running:

    print +(sort grep /[[:alpha:]]/, map { chr } 0..255), "\n";

To find out what [:alpha:] maps to.