Update the link to the NSIS wiki for the UnTGZ plugin.
Disambiguate the comment about the applicability of the built-in SSL files.

Mention the new multi-language support, the additional compiler
warnings about PFI_LANG_NSISDL_PLURAL and how to interpret them.

Change GPL license to remove the option for the person using POPFile
to license it under any version of the GPL other than v2