By default the UI operates in Stealth Mode and X-POPFile-Link uses "127.0.0.1". Some email clients (e.g. Eudora 6.x and 7.x) display an annoying warning/confirmation message when this link is clicked.
The new bayes_localhostname configuration item is only used when the
UI is operating in Stealth Mode.

By default the bayes_localhostname configuration item is empty
so X-POPFile-Link will use "127.0.0.1"

bayes_localhostname can be used to specify a name which stops these
messages from being displayed. Eudora still displays these messages
if "localhost" is used because it expects to find at least one dot
in the name (eg it is happy with a name like "froggie.gremlin")