Some more work on new Admin page.  Same issues as last commit, plus wrongly using Security_If_Local for disabling remote options while in Stealth Mode (that is for the UI only).  We have no global Stealth mode var so this is just a placeholder.
Either need a new PF var for the Stealth checkbox or loop through all the modules and if all are local we are in stealth mode.  With a sepearate Stealth mode var you could switch to Stealth Mode and leave your current Server Mode options selected.  They would be disabled checkboxes but still have checks in them.  But that could be confusing.

Added Logger Level drop down that needs connecting.