Updated to work with the NSIS 2.19 compiler (for the 0.22.5 release).
If no reconfigurable clients are detected the wizard no longer attempts
to find Outlook Express accounts (oops!).

If an Outlook Express account which uses SMTP Authentication is to be
reconfigured and it re-uses the POP3 username (the default setting) then
a separate SMTP username is created (to avoid login failures when SMTP
Authentication tries to use the POPFile-enabled POP3 username).