Updated to work with the NSIS 2.19 compiler (for the 0.22.5 release).
Switched from the standard NSISdl plugin to the Inetc plugin since it
offers much better proxy support and also better support for dial-up
connections. No longer need to ping the remote site to detect if the
target system is connected to the Internet.

Vpatch used when installing SSL Support files from the University of
Winnipeg repository. IO::Socket::SSL v0.99 and v0.999 are not compatible
with POPFile 0.22.x so this patch is used to downgrade the SSL.pm file
to v0.97 which is compatible.