Updated for the 0.22.5 release. SSL Support files are now downloaded by the Inetc plugin operating in NSISdl-compatible mode. The new plugin removes the need to ping the remote server to check the connection.
Both the installer and the SSL Setup wizard can patch IO::Socket::SSL
v0.99 or v0.999 to downgrade SSL.pm to v0.97 which is POPFile-compatible.

Outlook Express accounts which use SMTP Authorisation with the POP3
username are now given a separate SMTP username when they are configured
to work with POPFile.