IO::Socket::SSL v1.01 (released yesterday) is not compatible with the Windows version of POPFile 0.22.4 so the SSL Setup wizard has been updated to cope with this. Patch problems are now reported instead of simply being logged.
Updated to use the latest version of the Inetc plugin. By default the
install details are hidden and in order to avoid a strange looking display
the "Show Details" button is disabled whilst downloading the SSL files
(for some reason the new version of Inetc does not hide that button).