To make future maintenance easier MonitorCC.nsi now uses this library instead of having local copies of some of the library functions.
To permit non-admin users to correctly detect when POPFile or some of
its NSIS-based utilities are running, the LockedList plugin is used
instead of code which assumes the user has administrator rights. The
current version (v0.3 RC2) of this plugin requires some workarounds
to cope with the "unusual" mixed format path used for POPFile.