Previous releases of POPFile only supported one parser (Kakasi) for Japanese text. Update the installer to offer a choice of 3 parsers (Kakasi, MeCab and Internal) when Nihongo is selected. The MeCab files are too big to include in the installer so they will be downloaded during the installation if the MeCab option is selected.
The new Nihongo Parser pages and text are only shown when Nihongo is
selected (or if an English-only build is made) so string constants are
used instead of LanguageString commands (in order to keep things simple).

Removed support for the NO_KAKASI compile-time switch as it is no
longer relevant. (May introduce a similar switch affecting all of
the Nihongo Parser selection handling later on.)

Fix the bug in the NSIS2IO function that truncated MBCS strings.