Remove installer support for the 'Klingon' language.
Handle the following changes to the skins shipped with POPFile:

(a) use 'simplyblue' as the default skin

(b) add 'coolmint' and 'ocean' skins

(c) drop 'klingon', 'prjbluegrey' and 'prjsteelbeach' skins

(d) drop 'tinydefault' and 'lrclaptop' skins as they are almost
    identical to the 'smalldefault' skin

When upgrading an existing installation do not delete the folders
for the skins which have been dropped (because these skins will
more or less work in POPFile 1.0.0).

If 'klingon', 'prjbluegrey' or 'prjsteelbeach' is currently selected
then change popfile.cfg to use 'simplyblue' instead (the new default
skin for the 1.0.0 release)

If 'tinydefault' or 'lrclaptop' is currently selected then change
popfile.cfg to use the 'smalldefault' skin instead.