When upgrading an existing installation the installer will now pre-select the same optional POPFile components as the existing installation. The user is allowed to change this selection. (See ticket:26)
When upgrading a Japanese system the installer will pre-select
the user's current Nihongo parser. The user is allowed to change
this selection. (See ticket:13)

When the Add/Remove Program's "Change" button is used the user's
current Nihongo parser will be pre-selected and if the parser is
changed the user's popfile.cfg file will be updated with the new
value. (See ticket:14 and ticket:15)

The new minimal Perl and the PDK used to build the popfile*.exe
programs officially support only Windows 2000 or later. If an
older version of Windows is detected a warning is displayed.
(See ticket:104)

Fix minor bugs that resulted in some temporary files and folders
being left behind if the user cancelled the install (or uninstall)
operation. Found this while testing the above changes.