Add some code to show a useful compiler error message if the MD5 plugin is missing.
When replacing my system's hard disk I decided to re-install
Windows and the applications I use instead of simply cloning
the old hard disk. This is how I found a problem with the
utility which checks that the necessary extra NSIS plugins
are present: the utility cannot be compiled if one of these
extra plugins (the MD5 plugin, md5dll.dll) is missing!