Normally the POPFILE_USER environment variable has the same value as the HKCU registry data that records the location of the POPFile 'User Data' (the folder holding 'popfile.cfg').
However it is possible for POPFILE_USER to point to a different
set of POPFile 'User Data' (e.g. the installer creates a simple
batch file ('pfi-run.bat') that makes it easy to do this).

ticket:176 reported that the PFI Diagnostic Utility assumed
that if the environment variable was defined then it would
always point to the same location as the HKCU registry data.
These changes update the diagnostic utility to handle cases
where POPFILE_USER and the HKCU data point to different sets
of POPFile configuration data.