POPFile 1.1.4 will use Perl 5.16 which is not binary compatible with the Perl used by previous releases of POPFile. Start changing the Windows installer to cope with the new version of Perl.
Previous installers offered SOCKS V support as an option. However this
only involved copying one extra file to the minimal Perl so this file
is now included in the default minimal Perl.

The 'File' commands used to specify the SSL support have been moved
from 'installer.nsi' to this file to make it easier to maintain the
default minimal Perl.

Support for BerkeleyDB and SQLite 2.x format databases is not included
(because there are no readily available Perl 5.16 packages for these
old formats).